+++ src/server/loaders/imageLoader/imageLoader.ts
@@ -65,6 +65,9 @@ export const imageLoader: AssetLoader = async (
       compressionLevel: 9,
       loop: 0,
       delay: 100,
+      blurRadius: null,
+      rotate: null,
+      flip: null,
       ...defaultOptions,
       ...decodedQuery,
     } as TransformOptions;